Add files via upload
Oceanographers and Coastal engineers frequently require global bathymetry data that they can download from GEBCO - The General Bathymetric Chart of the Oceans. This script gets the bathymetry data from a GEBCO geotiff file, and the data are exported to two plain text files XYZ. In one of them, the coordinates are geographical, and in the second, the coordinates are UTM (WGS84).